refactor : 카테고리 타입 분리
